move managers and canvas_plotter to plotter module
