+++ pytripgui/canvas_vc/plotter/blit_manager.py
