+++ pytripgui/canvas_vc/plotter/placement_manager.py
